enable peerTest / multi-sim testing
